+++ src/main/java/com/example/tripKo/domain/address/entity/AddressCategory.java
@@ -1,13 +1,11 @@
 package com.example.tripKo.domain.address.entity;
 
-import com.example.tripKo.BaseTimeEntity;
+import com.example.tripKo.domain.BaseTimeEntity;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @DynamicUpdate
